### 2021-09-07 #3699 @whedon
[PRE REVIEW]: SPbLA: The Library of GPGPU-powered Sparse Boolean Linear Algebra Operations
**Submitting author:** @EgorOrachyov (<a href="http://orcid.org/0000-0002-0424-4059">Egor Orachev</a>)
**Repository:** <a href="https://github.com/JetBrains-Research/spbla" target ="_blank">https://github.com/JetBrains-Research/spbla</a>
**Version:** v1.0.0
