### 2021-09-07 #3699 @arfon
@EgorOrachyov – apologies, but we need you to convert your SVG images to PNGs. (I realize this will have worked fine with the GitHub Action). Could you do that please?
