### 2021-09-07 #3699 @EgorOrachyov
Yes, sure. I've already pushed figures converted to PNGs.

Best regards,
Egor Orachev

вт, 7 сент. 2021 г. в 14:50, Arfon Smith ***@***.***>:

> @EgorOrachyov <https://github.com/EgorOrachyov> – apologies, but we need
> you to convert your SVG images to PNGs. (I realize this will have worked
> fine with the GitHub Action). Could you do that please?
>
