### 2021-09-09 #3699 @Nikoleta-v3
@EgorOrachyov  thank you again for your submission with JOSS,  could you take a look through https://bit.ly/joss-reviewers to suggest some potential reviewers 😄
