### 2021-09-10 #3699 @Nikoleta-v3
👋 @hausen & @arghdos & @pszi1ard & @abb58 would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "SPbLA: The Library of GPGPU-powered Sparse Boolean Linear Algebra Operations". You can find more information at the top of this Github issue 🆙
