### 2021-09-16 #3699 @abb58
@Nikoleta-v3 Thanks for getting in touch. I am willing to participate in reviewing your submission to JOSS.
