### 2021-09-21 #3743 @whedon
[REVIEW]: SPbLA: The Library of GPGPU-powered Sparse Boolean Linear Algebra Operations
**Submitting author:** @EgorOrachyov (<a href="http://orcid.org/0000-0002-0424-4059">Egor Orachev</a>)
**Repository:** https://github.com/JetBrains-Research/spbla
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
