### 2021-09-21 #3743 @Nikoleta-v3
@EgorOrachyov please fix the missing DOI: https://github.com/openjournals/joss-reviews/issues/3743#issuecomment-923893328 :)
