### 2021-10-08 JetBrains-Research/spbla#9 @mlxd
Ensure CUDA architecture is set for targets
As part of [JOSS review 3743](https://github.com/openjournals/joss-reviews/issues/3743), I am following the build guide for the library. As of CMake 3.18, the following warning/error will be reported, as per https://cmake.org/cmake/help/latest/policy/CMP0104.html

This may cause issues on newer systems building the library.

```
CMake Warning (dev) in spbla/CMakeLists.txt:
  Policy CMP0104 is not set: CMAKE_CUDA_ARCHITECTURES now detected for NVCC,
  empty CUDA_ARCHITECTURES not allowed.  Run "cmake --help-policy CMP0104"
  for policy details.  Use the cmake_policy command to set the policy and
  suppress this warning.

  CUDA_ARCHITECTURES is empty for target "spbla".
This warning is for project developers.  Use -Wno-dev to suppress it.

CMake Warning (dev) in spbla/CMakeLists.txt:
  Policy CMP0104 is not set: CMAKE_CUDA_ARCHITECTURES now detected for NVCC,
  empty CUDA_ARCHITECTURES not allowed.  Run "cmake --help-policy CMP0104"
  for policy details.  Use the cmake_policy command to set the policy and
  suppress this warning.

  CUDA_ARCHITECTURES is empty for target "spbla".
This warning is for project developers.  Use -Wno-dev to suppress it.
```
