### 2021-11-21 JetBrains-Research/spbla#13 @mlxd
Additional build issues
Hi @egor-bogomolov I have attempted to continue building the repo but am hitting some issues still. For GCC 11, with CUDA 11.5 I made the following changes to ensure compilation. Mostly ensuring headers are included where needed. I have attached the diff below:

```
diff --git a/deps/nsparse/include/nsparse/unified_allocator.h b/deps/nsparse/include/nsparse/unified_allocator.h
index ea6a98b..10f36fc 100644
--- a/deps/nsparse/include/nsparse/unified_allocator.h
+++ b/deps/nsparse/include/nsparse/unified_allocator.h
@@ -3,7 +3,7 @@
 #include <thrust/detail/config.h>
 #include <thrust/device_ptr.h>
 #include <thrust/mr/allocator.h>
-#include <thrust/memory/detail/device_system_resource.h>
+//#include <thrust/memory/detail/device_system_resource.h>

 #include <limits>
 #include <stdexcept>
@@ -80,4 +80,4 @@ using managed = thrust::device_unified_allocator<T>;
 //template <typename T>
 //using managed_vector = thrust::device_vector<T, managed<T>>;

-}  // namespace nsparse
\ No newline at end of file
+}  // namespace nsparse
diff --git a/spbla/sources/cuda/cuda_instance.hpp b/spbla/sources/cuda/cuda_instance.hpp
index dd2b8fe..c5e3510 100644
--- a/spbla/sources/cuda/cuda_instance.hpp
+++ b/spbla/sources/cuda/cuda_instance.hpp
@@ -27,6 +27,8 @@

 #include <core/config.hpp>
 #include <unordered_set>
+#include <cstddef>
+

 namespace spbla {

@@ -69,4 +71,4 @@ namespace spbla {

 }

-#endif //SPBLA_CUDA_INSTANCE_HPP
\ No newline at end of file
+#endif //SPBLA_CUDA_INSTANCE_HPP
diff --git a/spbla/sources/sequential/sq_ewiseadd.hpp b/spbla/sources/sequential/sq_ewiseadd.hpp
index 616274f..0ca6c6a 100644
--- a/spbla/sources/sequential/sq_ewiseadd.hpp
+++ b/spbla/sources/sequential/sq_ewiseadd.hpp
@@ -25,6 +25,7 @@
 #ifndef SPBLA_SQ_EWISEADD_HPP
 #define SPBLA_SQ_EWISEADD_HPP

+#include <cstddef>
 #include <sequential/sq_csr_data.hpp>

 namespace spbla {
diff --git a/spbla/sources/sequential/sq_kronecker.hpp b/spbla/sources/sequential/sq_kronecker.hpp
index a65f0d2..7562942 100644
--- a/spbla/sources/sequential/sq_kronecker.hpp
+++ b/spbla/sources/sequential/sq_kronecker.hpp
@@ -25,6 +25,7 @@
 #ifndef SPBLA_SP_KRONECKER_HPP
 #define SPBLA_SP_KRONECKER_HPP

+#include <cstddef>
 #include <sequential/sq_csr_data.hpp>

 namespace spbla {
diff --git a/spbla/sources/sequential/sq_submatrix.hpp b/spbla/sources/sequential/sq_submatrix.hpp
index 62da1b1..02a0ac3 100644
--- a/spbla/sources/sequential/sq_submatrix.hpp
+++ b/spbla/sources/sequential/sq_submatrix.hpp
@@ -25,6 +25,7 @@
 #ifndef SPBLA_SQ_SUBMATRIX_HPP
 #define SPBLA_SQ_SUBMATRIX_HPP

+#include <cstddef>
 #include <sequential/sq_csr_data.hpp>

 namespace spbla {
diff --git a/spbla/sources/sequential/sq_transpose.hpp b/spbla/sources/sequential/sq_transpose.hpp
index 14b21ab..c85d584 100644
--- a/spbla/sources/sequential/sq_transpose.hpp
+++ b/spbla/sources/sequential/sq_transpose.hpp
@@ -26,6 +26,7 @@
 #define SPBLA_SQ_TRANSPOSE_HPP

 #include <sequential/sq_csr_data.hpp>
+#include <cstddef>

 namespace spbla {

diff --git a/spbla/sources/utils/csr_utils.hpp b/spbla/sources/utils/csr_utils.hpp
index ce509ea..82ff72f 100644
--- a/spbla/sources/utils/csr_utils.hpp
+++ b/spbla/sources/utils/csr_utils.hpp
@@ -27,6 +27,7 @@

 #include <core/config.hpp>
 #include <vector>
+#include <cstddef>

 namespace spbla {
```

After this, I can confirm 100% compilation, and I was able to run (a sample of) the tests.
