### 2021-11-21 #3743 @mlxd
Apologies for the delays. I have added a new issue for compiling the library at https://github.com/JetBrains-Research/spbla/issues/13
