### 2021-11-21 JetBrains-Research/spbla#14 @mlxd
Add additional examples to docs
To check off item `Example usage` in the review, some additional documented examples would be best. This should ideally showcase the various available functionalities in the library.
