### 2021-11-21 #3743 @mlxd
Issue added for adding additional examples: https://github.com/JetBrains-Research/spbla/issues/14
