### 2021-11-21 JetBrains-Research/spbla#15 @mlxd
Generate API of Python functionality
The Python API, assuming sufficiently listed docstrings, can be used to autogenerate a documented API. This would assist with usage of the library, and be complimentary to the example usage request #14
