### 2021-11-21 JetBrains-Research/spbla#16 @mlxd
Add community guidelines for contribution
To tick off item `Community guidelines` in the review adding contribution guidelines to the repository would help. This can be a paragraph in the readme, as well as templated issue/PR items for user submissions.
