### 2021-11-21 #3743 @mlxd
Issue added for community guidelines https://github.com/JetBrains-Research/spbla/issues/16
