### 2021-11-21 JetBrains-Research/spbla#15 @mlxd
In addition, the C++ documentation is not satisfactorily documented for end-users. If the purpose is to only expose the Python API, then this should also be clearly stated.
