### 2021-11-21 #3743 @mlxd
Issue added for API functionality documentation https://github.com/JetBrains-Research/spbla/issues/15
