### 2021-11-22 JetBrains-Research/spbla#15 @EgorOrachyov
> In addition, the C++ documentation is not satisfactorily documented for end-users. If the purpose is to only expose the Python API, then this should also be clearly stated.

@mlxd Hi, yes. C++ sources of the library are not exposed to the end (not the developer of the library) user.
C API and python package is the only way to use library primitives. So, will it be enough if we state the list of available APIs in the project readme?
