### 2021-11-22 JetBrains-Research/spbla#15 @mlxd
Hi @EgorOrachyov I think a listing of the public-facing C-API (function names & arguments) alongside a quick description/docstring for each should work. I fear listing the function-name alone is not sufficient documentation.

Same again for the Python functionality. Any expected public-facing API or module should have an included docstring that can be user-accessible.
