### 2022-04-20 #3743 @Nikoleta-v3
👋 @hausen & @arghdos would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "SPbLA: The Library of GPGPU-powered Sparse Boolean Linear Algebra Operations". You can find more information at the top of this Github issue 🆙
