### 2022-05-15 #3743 @Nikoleta-v3
👋 @bencardoen & @chaypaterson & @alejandrogallo & @mkareshk  would any of you be willing to review this submission for JOSS? One reviewer has already reviewed the submission, and I am looking for a second pair of 👀

The submission I'd like you to review is titled: "SPbLA: The Library of GPGPU-powered Sparse Boolean Linear Algebra Operations". You can find more information at the top of this Github issue 🆙
