### 2022-05-16 #3743 @bencardoen
@Nikoleta-v3 the invite / confirm link at the top of the issue seems no longer valid, how do I start the review? (this is my first JOSS review)
