### 2022-05-16 #3743 @Nikoleta-v3
Thank you very much @bencardoen for agreeing to review this submission! 🙏🏻

> this is my first JOSS review

This is great - welcome! 🎉

If you comment the following command `@editorialbot generate my checklist` on this issue the bot will generate a checklist for you ✍🏻

What you have to do is go over the checklist and make sure that the submission addresses each point. These points will include details both about the software and the paper.

I recommend that each time you would like the authors to change something that you open an issue on their repository (https://github.com/JetBrains-Research/spbla) and in there mention this issue (#3743). That way I can keep track of the issues!

Once you are done with the checklist you can just leave a comment here to let me know. If you have any questions regarding something on the checklist (or anything else) you can ping me here and I will reply as soon as possible!
