### 2022-05-21 JetBrains-Research/spbla#17 @bencardoen
Building / tests fail on missing Cmake in gtest
In ref to review https://github.com/openjournals/joss-reviews/issues/3743
**Describe the bug**
Building/executing gtest fails, perhaps I missed something in the build instructions ?

**To Reproduce**
(after git clone, note I set CUDA to OFF in CMakeLists.txt
```bash
❯ mkdir build
❯ cd build
❯ ls
❯ cmake ..
-- The CXX compiler identification is GNU 10.3.1
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
-- Build spbla in release mode (default: was not specified)
-- The C compiler identification is GNU 10.3.1
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Looking for CL_VERSION_2_2
-- Looking for CL_VERSION_2_2 - not found
-- Looking for CL_VERSION_2_1
-- Looking for CL_VERSION_2_1 - not found
-- Looking for CL_VERSION_2_0
-- Looking for CL_VERSION_2_0 - not found
-- Looking for CL_VERSION_1_2
-- Looking for CL_VERSION_1_2 - found
-- Found OpenCL: /usr/local/cuda/lib64/libOpenCL.so (found suitable version "1.2", minimum required is "1.1")
-- Add googletest as unit-testing library
CMake Error at CMakeLists.txt:58 (add_subdirectory):
  The source directory

    /home/bcardoen/SFUVault/spbla/deps/gtest

  does not contain a CMakeLists.txt file.

-- Add CPU sequential fallback backend
-- Add OpenCL backend for GPGPU computations
-- Add unit tests directory to the project
-- Configuring incomplete, errors occurred!
See also "/home/bcardoen/SFUVault/spbla/build/CMakeFiles/CMakeOutput.log".
See also "/home/bcardoen/SFUVault/spbla/build/CMakeFiles/CMakeError.log".
```

**Expected behavior**
Build success

**Environment**
- OS name and version: Fedora 33
- GPU name and vendor: Nvidia GeForce 1050Ti
- GPU driver version Driver Version: 460.56       CUDA Version: 11.2

**Build Configuration**
- Compiler version
```
Using built-in specs.
COLLECT_GCC=gcc
COLLECT_LTO_WRAPPER=/usr/libexec/gcc/x86_64-redhat-linux/10/lto-wrapper
OFFLOAD_TARGET_NAMES=nvptx-none
OFFLOAD_TARGET_DEFAULT=1
Target: x86_64-redhat-linux
Configured with: ../configure --enable-bootstrap --enable-languages=c,c++,fortran,objc,obj-c++,ada,go,d,lto --prefix=/usr --mandir=/usr/share/man --infodir=/usr/share/info --with-bugurl=http://bugzilla.redhat.com/bugzilla --enable-shared --enable-threads=posix --enable-checking=release --enable-multilib --with-system-zlib --enable-__cxa_atexit --disable-libunwind-exceptions --enable-gnu-unique-object --enable-linker-build-id --with-gcc-major-version-only --with-linker-hash-style=gnu --enable-plugin --enable-initfini-array --with-isl --enable-offload-targets=nvptx-none --without-cuda-driver --enable-gnu-indirect-function --enable-cet --with-tune=generic --with-arch_32=i686 --build=x86_64-redhat-linux
Thread model: posix
Supported LTO compression algorithms: zlib zstd
gcc version 10.3.1 20210422 (Red Hat 10.3.1-1) (GCC)

```
- SDK version
- CMake version cmake version 3.19.7
