### 2022-05-21 JetBrains-Research/spbla#18 @bencardoen
Document first invocation of API is slower
In ref to review https://github.com/openjournals/joss-reviews/issues/3743
(minor issue)
The first time I execute
```
import pyspbla as sp
b = sp.Matrix.empty(shape=(3, 4))
```
there's a 1-2 sec wait before results show up (which I guess is due to first call init), it can help document this (unless I overlooked this).
