### 2022-05-21 JetBrains-Research/spbla#19 @bencardoen
Paper feedback
In ref to review https://github.com/openjournals/joss-reviews/issues/3743
Summary feedback on the manuscript

First off, overall the paper is well written, API/implementation is well described, well documented library/API.

This feedback is in 3 parts, writing (minor to no issues),  and a question on the memory cost of addition, evaluation (some comments).

## Writing
Introduction:
I'd rewrite the first sentence into something like:
```
Answering research questions in data analysis often involves expressing the solution in terms of matrix/vector operations.
```
I understand what you're saying in its current form, just concerned you're underselling the contribution.

```
This way it is possible to employ a set  -->  ... to leverage a set of powerful ...
```

Section IV
```
applicability --> utility
```

## Method
End of section III
```
"it can negatively affect the memory consumption for large matrices with lots of duplicated non-zero values at the same positions."
```
Can this not be done without extra allocation by doing a sparse vector elementwise multiplication, reusing your own API?
It could be of interest to review (for future work, not this review necessarily)

## Evaluation
###  Table II+III
- If you report mean of 10 (or in general k) runs, please include variance and/or confidence intervals. Means can obscure interesting effects that can be exploited to optimize the API.
- Consider that arithmetic mean is not always appropriate in performance/standardized benchmarks, see Fleming's seminal paper for background[1]. That being said, I do understand that you're bound to how others report on standard benchmark datasets, I don't expect you to redo/review the current results/experimetns, this is a suggestion only for future analysis.
- Please boldface best / worst results, humans take a long time to parse tabular numerical data, helping with highlights makes it faster to draw conclusions from the tabular data
- Consider using a lineplot / regression (LOESS) with mean + confidence interval, this would clearly show differences and patterns, and highlight the difference in performance
- Did you consider effect of hot/cold cache, e.g. my assumption would be the first invocation would be the slowest if that is a factor, did you see the same?

Thanks,

Ben

[1] https://doi.org/10.1145/5666.5673
