### 2022-05-21 JetBrains-Research/spbla#17 @bencardoen
Forgot to link https://github.com/openjournals/joss-reviews/issues/3743
