### 2022-05-21 #3743 @bencardoen
@Nikoleta-v3
I completed the checklist, and made issues (I hope in the right way) for the items I came across.
My general review comments are in issue https://github.com/JetBrains-Research/spbla/issues/19
I haven't been able to verify the performance results due to a build issue https://github.com/JetBrains-Research/spbla/issues/17 so that's unchecked, and I still need to get the python tests running.
I'll revisit this comment thread once I've completed the Python part.
