### 2022-05-25 JetBrains-Research/spbla#18 @EgorOrachyov
@bencardoen Hi, thank you for your feedback.

Is it really necessary to document such a behavior, since:
- module import requires library initialization
- it is a common practice to do a `warm up`  before the execution or measurements
- first invocation may require kernels compilations
- on Cuda backend, if driver already cached native kernels from ptx assembly, first invocation is not necessary slow

So, what do you think?
