### 2022-05-30 JetBrains-Research/spbla#18 @bencardoen
Hi @EgorOrachyov,
You don't need to, it's a suggestion/minor comment only. I figured adding it to the docs can be of help to e.g. algorithm designers who want to use your API but are not versed in the background implementation.

I came across it when trying to build a singularity image, adding the import to the definition file ensures the image contains the prebuilt version, rather than having to add it to code that depends on the image.
