### 2022-06-14 #3743 @bencardoen
@Nikoleta-v3 https://github.com/JetBrains-Research/spbla/issues/19 (general feedback/comments) is open, the other issues have been addressed. I've updated the comment above to reflect the changes.
