### 2022-06-29 #3743 @EgorOrachyov
@Nikoleta-v3 Hi! Yes, we have addressed some minor technical and documentation related features. Now we are going to do a text clean-up, as it is suggested in https://github.com/JetBrains-Research/spbla/issues/19 issue. I hope, we will have it done in a couple of days.

Apologies for delays, busy work days :/
