### 2022-07-03 JetBrains-Research/spbla#19 @EgorOrachyov
@bencardoen Hi, thank you for your question.

> "it can negatively affect the memory consumption for large matrices with lots of duplicated non-zero values at the same positions."
> Can this not be done without extra allocation by doing a sparse vector elementwise multiplication, reusing your own API?

The problem is that the description is not clear enough. COO matrices element-wise addition based on merge-path effectively does the following thing: 1) sorted merge and then 2) reduction of duplicates. For the first step we must allocate temporary memory with worst case scenario `nnz(A) + nnz(B)`. Thus, thus approach degrades significantly, if we decide to add equal matrices. In this case we allocate `2*nnz(A)` memory, however result matrix `C = A + A` takes only `nnz(A)` memory. You can imagine, if matrix `A` takes nearly 2 GiB of VRAM, such addition will eat 4 GiB of temporary space, what is an overkill.

In order to avoid such a behavior, utilization of more fancy and memory friendly algorithms is required.
