### 2022-07-03 JetBrains-Research/spbla#19 @EgorOrachyov
@bencardoen Hi thank you for your feedback. I hope, we have addressed the issue in the updated paper text https://github.com/JetBrains-Research/spbla/actions/runs/2604632019

* Fixed some writing in introduction, evaluation. Add more info on experiment (mentioned hot/cold start)
* Address method question
* Evaluation: highlight best/worst cases, add minor info about variation of results + hot/cold start. My apologies, we are currently not able to remake plots to fully address the issue.
