### 2022-07-03 #3743 @EgorOrachyov
@Nikoleta-v3 Hi, I hope, we have addressed all issues. Waiting for a feedback from @bencardoen

It would be great, if you trigger paper recompilation (seems I do not have enough permissions to do it), since we updates text and some diagrams :)

`@whedon generate pdf from branch joss-paper`
