### 2022-07-04 #3743 @EgorOrachyov
@bencardoen Hi, thank you once more 😀! I've minor fixed writing issues, new pdf file is above.
