### 2022-07-09 #3743 @Nikoleta-v3
@EgorOrachyov thank you for your patience. There are a few typos in the paper. I believe I've picked them all up (see below).

Another minor comment I have is regarding the figures. The numbers (the memory consumption) are not always easy to see. Could you maybe use a bold font?

### Typos
#### Summary

- as **a** stand-alone self-sufficient

#### Statement of need

-  to leverage a set of ~a~ powerful sparse linear algebra libraries (I think)
- Also, we hope ~,~ that the library is a small step to move the implementation (remove second comma)

#### Related tools

- is one of the first standards ~,~ that formalize the mathematical (remove comma)
- in the language of ~the~ linear algebra.
- is a reference implementation~s~ of the (I think)
- It is **a** mature and fully featured library with **a** number of bindings for
- is currently in ~the~ active development.
- There **are a** number of open-source and proprietary libraries

#### Performance

- The experiment itself is designed as a computational task~s~,  which arises as **a** stand-alone
- memory consumption in some cases, ~what~ **which** can be a critical limitation

#### Future research

- in order to support arbitrary types and operations, defined by **the** user.
