### 2022-08-19 #3743 @EgorOrachyov
@Nikoleta-v3

Hello! I apologize for such a long response.
We have corrected typos and updated the charts, highlighting the measurements numbers in bold.

I hope to proceed without delay to carry out the rest of the work 👽
