### 2022-08-19 #3743 @EgorOrachyov
@Nikoleta-v3

- [x] Release `v1.0.0` is ok. (I have made new one especially for Zenodo `v1.0.0-zenodo`, but it is effectively the same thing)
- [x] Zenodo archive [DOI: 10.5281/zenodo.7009938](https://doi.org/10.5281/zenodo.7009938)
- [x] Meta filled, orcid added where possible
- [x] [DOI: 10.5281/zenodo.7009938](https://doi.org/10.5281/zenodo.7009938)
