### 2022-08-19 #3743 @EgorOrachyov
@Nikoleta-v3 Seems to be fixed now

> @EgorOrachyov there is a missing DOI. Could you please take care of this? 😄
