### 2022-08-19 #3743 @Nikoleta-v3
dear editors please note that I've set the version to be the tagged release `v1.0.0`. However,

> Release v1.0.0 is ok. (I have made new one especially for Zenodo v1.0.0-zenodo, but it is effectively the same thing)

So the Zenodo archive is based on the tagged release `v1.0.0-zenodo`.  Let me know if I should change the version to be `v1.0.0-zenodo` instead.
