### 2022-08-21 #3743 @bencardoen
Congrats on the acceptance @EgorOrachyov
As a general comment, 99% of journals/venues I've interacted with so far could learn a lot of how the review process works here: highly responsive, collaborative, constructive, open, automated where needed, allowing humans to focus on the parts that need human reflection. @arfon @Nikoleta-v3
